+++ spec/controllers/enforce_immediate_filter_spec.rb
@@ -43,7 +43,7 @@ describe DummyImmediateController, type: :controller do
       $self = nil
       $arg = nil
       with_bot_challenge_config(BotChallengePage::BotChallengePageController,
-        after_challenge: ->(bot_detect_class) {
+        after_blocked: ->(bot_detect_class) {
           $triggered = true;
           $self = self;
           $arg = bot_detect_class